+++ tests/test_etc.py
@@ -13,7 +13,7 @@ def test_teardown_without_setup():
 
 @pytest.mark.skipif(sys.version_info < (3, 0), reason='whole project test are for Py 3')
 def test_whole_project_testing(request):
-    configure_djangoapp_plugin('pytest_djangoapp.tests.settings_project')
+    configure_djangoapp_plugin('tests.settings_project')
     settings_dict = Configuration.get_combined(request.config)
     assert '_' not in settings_dict
     assert 'SOME_ATTR' in settings_dict