+++ tests/test_fixtures_utils.py
@@ -1,4 +1,4 @@
 
 def test_conf_app_name(conf_app_name):
 
-    assert conf_app_name == 'pytest_djangoapp'
+    assert conf_app_name == 'tests'